python/requirements.txt
onnxruntime==1.18.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
